docs: improve documentation and scripts, provide script as executables
Improve the already useful documentation and factor out scripts as actual
shell scripts that the user can execute, this is handier and less error
prone.

The documentation provides direct links to the script to make it easier
to download them when following the instructions.

The instructions now use sudo only when it's strictly necessary and
avoid copying files outside of the user home directory as much as
possible.